[Certif] BSR Split acceptance tests into 3 files
